lab2 complete.

+++ lab2/DebugPractice/DebugExercise1.java
@@ -2,11 +2,12 @@
  * Exercise for learning how the debug, breakpoint, and step-into
  * feature work.
  */
+import java.lang.Math;
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
-        return result;
+        double quotient = (double) top / (double) bottom;
+        double result = Math.round(quotient);
+        return (int) result;
     }
 
     public static void main(String[] args) {
